Move drivetrain objects
Moved all drivetrain related classes (drivetrain class and related
states) to follow the following folder structure
subsystems:
  subsystem named folder:
    subsystem class
    states folder:
      all states for that subsystem
