+++ src/main/java/frc/robot/RobotContainer.java
@@ -15,11 +15,11 @@ import frc.robot.commands.ArmSetpointCommand;
 import frc.robot.commands.ChargeCircleCommand;
 import frc.robot.commands.ConeGrabState;
 import frc.robot.commands.CubeGrabState;
-import frc.robot.commands.DriveState;
 import frc.robot.constants.TelemetryConstants.Limelights;
 import frc.robot.subsystems.ArmSubsystem;
 import frc.robot.subsystems.ClawSubsystem;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.drivetrain.Drivetrain;
+import frc.robot.subsystems.drivetrain.states.DriveState;
 import frc.robot.utils.XboxController;
 import frc.team4272.controllers.utilities.JoystickAxes;
 import frc.team4272.controllers.utilities.JoystickAxes.DeadzoneMode;
