+++ src/main/java/frc/robot/commands/AprilRunCommand.java
@@ -1,7 +1,8 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.drivetrain.Drivetrain;
+import frc.robot.subsystems.drivetrain.states.PathFollowState;
 
 import static frc.robot.constants.AutoConstants.Paths.getGlobalTrajectories;
