+++ src/main/java/frc/robot/commands/ChargeCircleCommand.java
@@ -2,7 +2,8 @@ package frc.robot.commands;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.drivetrain.Drivetrain;
+import frc.robot.subsystems.drivetrain.states.PathFollowState;
 
 import static frc.robot.constants.AutoConstants.Paths.getGlobalTrajectories;
